requirements.txt
streamlit
numpy
pydicom
Pillow
python-docx
google-generativeai
PyMuPDF
requests

